Dequeue index & add support for sqlite RETURNING clause (#24)
* upgrade gorqlite to v0.0.9 (rqlite to v7.20.4)
* add support for sqlite RETURNING clause and use it in dequeue
* asynq tasks
   * add index (queue_name,state) on table asynq_tasks for use in dequeue
* completed tasks
   * move completed tasks to their own table
   * indexes: task_uuid,  (queue_name,retain_until) 
   * unique constraint: 'task_uuid'
* schema
   * refactor schema handling
   * always create tables (if not exist) instead of relying on the existence of a single one
* perf/debug: add logs on sqlite3.SQLiteConnection (and configuration for sqlite tracing)
* tests
   * internal/rqlite tests: don't remove sqlite file when test failed
   * add sqlite3 tests to run_all_tests.sh
* fix nil checks in NewErrorMux & NewRetryMux

+++ internal/rqlite/rqlite_setup_test.go
@@ -71,7 +71,11 @@ func setup(tb testing.TB) *RQLite {
 	})
 	if sqliteDbTemp {
 		tb.Cleanup(func() {
-			_ = os.Remove(config.SqliteDbPath)
+			if !tb.Failed() {
+				_ = os.Remove(config.SqliteDbPath)
+			} else {
+				fmt.Println("preserving sqlite DB at:", config.SqliteDbPath)
+			}
 		})
 	}
